refactor: [#1191] extract mod auth::key
